fixup! KTOR-8193 Introduce non-blocking i/o stream wrappers for channels
